refactor: 메시지 dto 패키지 변경
